Add documentation and UI enhancements
Added comprehensive documentation files covering architecture, database schema, cycle update recalculation, consumption card, and create consumption enhancements. Updated README with project overview, features, and setup instructions. Introduced new dashboard components (consumption card, cycle picker strip), improved number formatting, and enhanced create/edit flows for cycles and consumptions. Migrated dashboard logic from Bloc to Riverpod, removed obsolete Bloc files, and refactored related UI and repository code for better maintainability.

+++ lib/domain/repositories/electricity_readings_repository.dart
@@ -12,16 +12,16 @@ abstract class ElectricityReadingsRepository {
     required String houseId,
     required String cycleId,
     required DateTime date,
-    required int meterReading,
-    required int unitsConsumed,
+    required double meterReading,
+    required double unitsConsumed,
     required double totalCost,
     String? notes,
   });
   Future<void> updateReading({
     required String id,
     DateTime? date,
-    int? meterReading,
-    int? unitsConsumed,
+    double? meterReading,
+    double? unitsConsumed,
     double? totalCost,
     String? notes,
   });
